backend/requirements.txt
fastapi
uvicorn[standard]
pybaseball
pandas
numpy
statsmodels
scipy
python-dotenv
supabase
PyJWT
stripe
sendgrid
apscheduler
jinja2
psycopg2-binary
